requirements.txt
Flask
Flask-Cors
requests
logger
PyJWT
urllib3
psycopg2-binary